Removing ServerBootstrapFactory.
