+++ wildfly-module/src/main/resources/org/jboss/aerogear/simplepush/subsystem/LocalDescriptions.properties
@@ -5,7 +5,6 @@ simplepush.server=SimplePush Server type
 simplepush.server.add=Operation Adds SimplePush server 
 simplepush.server.remove=Operation Removes SimplePush server 
 simplepush.server.socket-binding=Socket binding attribute for SimplePush server
-simplepush.server.factory-class=SimplePush ServerBootstrap factory attribute
 simplepush.server.thread-factory=The thread-factory to be used by Netty
 simplepush.server.datasource-jndi-name=An optional datasource JNDI name that a service can depend on
 simplepush.server.token-key=An token key used for encryption/decryption
