--- wildfly-module/src/test/java/org/jboss/aerogear/simplepush/subsystem/MockServerBootstrapFactory.java
@@ -1,51 +1,0 @@
-/**
- * JBoss, Home of Professional Open Source
- * Copyright Red Hat, Inc., and individual contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.aerogear.simplepush.subsystem;
-
-import java.util.concurrent.ThreadFactory;
-
-import io.netty.bootstrap.ServerBootstrap;
-import io.netty.channel.ChannelInitializer;
-import io.netty.channel.EventLoopGroup;
-import io.netty.channel.nio.NioEventLoopGroup;
-import io.netty.channel.socket.SocketChannel;
-import io.netty.channel.socket.nio.NioServerSocketChannel;
-
-import org.jboss.as.network.SocketBinding;
-
-public class MockServerBootstrapFactory implements ServerBootstrapFactory {
-
-    @Override
-    public ServerBootstrap createServerBootstrap(final SocketBinding socketBinding,
-            final ThreadFactory threadFactory,
-            final String tokenKey,
-            final boolean endpointTls) {
-        final EventLoopGroup bossGroup = new NioEventLoopGroup();
-        final EventLoopGroup workerGroup = new NioEventLoopGroup();
-        final ServerBootstrap sb = new ServerBootstrap();
-        sb.group(bossGroup, workerGroup)
-        .channel(NioServerSocketChannel.class)
-        .childHandler(new ChannelInitializer<SocketChannel>() {
-            @Override
-            protected void initChannel(SocketChannel ch) throws Exception {
-            }
-        });
-        return sb;
-    }
-
-}
